# swagger-documentation.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Registration Successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/verify-email/{token}":{"post":{"operationId":"AuthController_confirmEmail","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/resend-email-verification-code":{"post":{"operationId":"AuthController_resendEmailVerificationCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/resend-forgot-password":{"post":{"operationId":"AuthController_resendForgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/reset-password":{"patch":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/change-password":{"patch":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/logout":{"get":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/users":{"get":{"operationId":"UsersController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Users retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["Users"]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Review deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Users"]}},"/api/v1/users/me":{"get":{"operationId":"UsersController_findMe","parameters":[],"responses":{"200":{"description":"user retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]}},"/api/v1/users/upload":{"put":{"operationId":"UsersController_UploadUserPhoto","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"photo":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]}},"/api/v1/reviews":{"post":{"operationId":"ReviewsController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new review","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":"review created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}}},"tags":["Reviews"]},"get":{"operationId":"ReviewsController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Reviews retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewDto"}}}}}},"tags":["Reviews"]}},"/api/v1/reviews/{id}":{"get":{"operationId":"ReviewsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Review retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}}},"tags":["Reviews"]},"patch":{"operationId":"ReviewsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReviewsDto"}}}},"responses":{"200":{"description":"Review Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDto"}}}}},"tags":["Reviews"]},"delete":{"operationId":"ReviewsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Review deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Reviews"]}},"/api/v1/product":{"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new product","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"product created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}},"tags":["Product"]},"get":{"operationId":"ProductController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Products retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Paginate"}}}}},"tags":["Product"]}},"/api/v1/product/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Product retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}},"tags":["Product"]},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"Product Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}},"tags":["Product"]},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Product deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Product"]}},"/api/v1/brand":{"post":{"operationId":"BrandController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new brand","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBrandDto"}}}},"responses":{"201":{"description":"brand created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandDto"}}}}},"tags":["Brand"]},"get":{"operationId":"BrandController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Brand retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BrandDto"}}}}}},"tags":["Brand"]}},"/api/v1/brand/{id}":{"get":{"operationId":"BrandController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Brand retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandDto"}}}}},"tags":["Brand"]},"patch":{"operationId":"BrandController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBrandDto"}}}},"responses":{"200":{"description":"Brand Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandDto"}}}}},"tags":["Brand"]},"delete":{"operationId":"BrandController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Brand deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Brand"]}},"/api/v1/order":{"post":{"operationId":"OrderController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new order","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":"order created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}}},"tags":["Order"]},"get":{"operationId":"OrderController_findAll","parameters":[],"responses":{"200":{"description":"Order retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDto"}}}}}},"tags":["Order"]}},"/api/v1/order/{id}":{"get":{"operationId":"OrderController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Order retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}}},"tags":["Order"]},"patch":{"operationId":"OrderController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":"Order Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}}},"tags":["Order"]},"delete":{"operationId":"OrderController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Order deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Order"]}},"/api/v1/order/user/me":{"get":{"operationId":"OrderController_findUserOrders","parameters":[],"responses":{"200":{"description":"Order retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}}},"tags":["Order"]}},"/api/v1/category":{"post":{"operationId":"CategoryController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new category","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"category created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}},"tags":["Category"]},"get":{"operationId":"CategoryController_findAll","parameters":[],"responses":{"200":{"description":"Category retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}}}}}},"tags":["Category"]}},"/api/v1/category/{id}":{"get":{"operationId":"CategoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Category retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}},"tags":["Category"]},"patch":{"operationId":"CategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":"Category Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}},"tags":["Category"]},"delete":{"operationId":"CategoryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Category deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Category"]}},"/api/v1/subCategory":{"post":{"operationId":"SubCategoryController_create","parameters":[],"requestBody":{"required":true,"description":"Create a new subCategory","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubCategoryDto"}}}},"responses":{"201":{"description":"subCategory created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubCategoryDto"}}}}},"tags":["Sub-Category"]},"get":{"operationId":"SubCategoryController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","schema":{"type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"SubCategory retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SubCategoryDto"}}}}}},"tags":["Sub-Category"]}},"/api/v1/subCategory/{id}":{"get":{"operationId":"SubCategoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"SubCategory retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubCategoryDto"}}}}},"tags":["Sub-Category"]},"patch":{"operationId":"SubCategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSubCategoryDto"}}}},"responses":{"200":{"description":"SubCategory Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubCategoryDto"}}}}},"tags":["Sub-Category"]},"delete":{"operationId":"SubCategoryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"SubCategory deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Sub-Category"]}}},"info":{"title":"NestJs Hackathon Kit","description":"NestJs API Starter Hackthon Kit","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8989"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"password":{"type":"string"},"passwordConfirm":{"type":"string"}},"required":["name","email","phoneNumber","password","passwordConfirm"]},"User":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"reachme@amitavroy.com"},"password":{"type":"string","description":"Password in plain text","example":"Password@123"}},"required":["email","password"]},"EmailDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"reachme@amitavroy.com"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"password":{"type":"string","description":"Password in plain text","example":"Password@123"},"passwordConfirm":{"type":"string","description":"Password in plain text","example":"Password@123"},"resetToken":{"type":"string","description":"Reset Code sent to Emailt","example":"4972fe"}},"required":["password","passwordConfirm","resetToken"]},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string","description":"Password in plain text","example":"Password@123"},"newPassword":{"type":"string","description":"Password in plain text","example":"Password@123"},"newPasswordConfirm":{"type":"string","description":"Password in plain text","example":"Password@123"}},"required":["oldPassword","newPassword","newPasswordConfirm"]},"UpdateUserDto":{"type":"object","properties":{}},"StandardResponse":{"type":"object","properties":{"status":{"type":"string"},"code":{"type":"number"},"message":{"type":"string"},"data":{"type":"object"},"meta":{"type":"object"}},"required":["status","code","message","data","meta"]},"CreateReviewDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"rating":{"type":"number","minimum":0,"maximum":5},"productId":{"type":"string"}},"required":["title","description","rating","productId"]},"Brand":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"Category":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"SubCategory":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"ProductDto":{"type":"object","properties":{"id":{"type":"string"},"new":{"type":"boolean"},"name":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"},"features":{"type":"string"},"price":{"type":"number"},"inStock":{"type":"boolean"},"quantityInStock":{"type":"number"},"coverImage":{"type":"string"},"imageGallery":{"type":"object","example":{"image1":"https://dog-image-1.jpg","image2":"https://dog-image-2.jpg"}},"metaData":{"type":"object","example":{"specifications":{"ram":"6gb","rom":"128gb"},"includedItems":[{"quantity":1,"item":"phone case"}]}},"averageRating":{"type":"number"},"noOfRatings":{"type":"number"},"reviews":{"type":"array","items":{"type":"string"}},"brand":{"$ref":"#/components/schemas/Brand"},"category":{"$ref":"#/components/schemas/Category"},"subCategory":{"$ref":"#/components/schemas/SubCategory"}},"required":["id","new","name","slug","description","features","price","inStock","quantityInStock","coverImage","imageGallery","metaData","averageRating","noOfRatings","reviews","brand","category","subCategory"]},"ReviewDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"rating":{"type":"number","minimum":0,"maximum":5},"user":{"$ref":"#/components/schemas/User"},"product":{"$ref":"#/components/schemas/ProductDto"}},"required":["title","description","rating","user","product"]},"UpdateReviewsDto":{"type":"object","properties":{}},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"features":{"type":"string"},"price":{"type":"number"},"quantityInStock":{"type":"number"},"coverImage":{"type":"string"},"imageGallery":{"type":"object","example":{"image1":"https://dog-image-1.jpg","image2":"https://dog-image-2.jpg"}},"metaData":{"type":"object","description":"Additional Product Information","example":{"specifications":{"ram":"6gb","rom":"128gb"},"includedItems":[{"quantity":1,"item":"phone case"}]}}},"required":["name","description","features","price","quantityInStock","coverImage","imageGallery","metaData"]},"meta":{"type":"object","properties":{"count":{"type":"number"},"currentPage":{"type":"number"},"nextPage":{"type":"number"},"prevPage":{"type":"number"},"lastPage":{"type":"number"}},"required":["count","currentPage","nextPage","prevPage","lastPage"]},"Paginate":{"type":"object","properties":{"status":{"type":"string"},"data":{"type":"array","items":{"type":"string"}},"meta":{"$ref":"#/components/schemas/meta"}},"required":["status","data","meta"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"features":{"type":"string"},"price":{"type":"number"},"quantityInStock":{"type":"number"},"coverImage":{"type":"string"},"imageGallery":{"type":"object","example":{"image1":"https://dog-image-1.jpg","image2":"https://dog-image-2.jpg"}},"metaData":{"type":"object","description":"Additional Product Information","example":{"specifications":{"ram":"6gb","rom":"128gb"},"includedItems":[{"quantity":1,"item":"phone case"}]}}}},"CreateBrandDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"BrandDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}},"required":["id","name","description"]},"UpdateBrandDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"ShippingInfo":{"type":"object","properties":{"address":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"zipCode":{"type":"string"}},"required":["address","city","country","zipCode"]},"CreateOrderDto":{"type":"object","properties":{"shippingInfo":{"$ref":"#/components/schemas/ShippingInfo"},"cartItems":{"example":[null],"type":"array","items":{"type":"string"}}},"required":["shippingInfo","cartItems"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"photo":{"type":"string"},"isEmailVerified":{"type":"boolean"},"role":{"type":"string","enum":["user","manager","admin"]}},"required":["id","name","email","phoneNumber","photo","isEmailVerified","role"]},"OrderDto":{"type":"object","properties":{"id":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"},"shippingInfo":{"allOf":[{"$ref":"#/components/schemas/ShippingInfo"}]},"orderStatus":{"type":"string","enum":["Placed","Cancelled","Confirmed","Shipped","Delivered"],"default":"Placed"},"total":{"type":"number"},"shippingFee":{"type":"number"},"grandTotal":{"type":"number"},"cartItems":{"example":[null],"type":"array","items":{"type":"string"}}},"required":["id","user","shippingInfo","orderStatus","total","shippingFee","grandTotal","cartItems"]},"UpdateOrderDto":{"type":"object","properties":{"shippingInfo":{"$ref":"#/components/schemas/ShippingInfo"},"cartItems":{"example":[null],"type":"array","items":{"type":"string"}}}},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"CategoryDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"childrenCategories":{"type":"array","items":{"type":"string"}}},"required":["id","name","description","childrenCategories"]},"UpdateCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"CreateSubCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"parentCategoryId":{"type":"string"}},"required":["name","description","parentCategoryId"]},"SubCategoryDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"parentCategory":{"$ref":"#/components/schemas/Category"}},"required":["id","name","description","parentCategory"]},"UpdateSubCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"parentCategoryId":{"type":"string"}}}}}}